pyproject.toml
[tool.poetry]
name = "pipcpay-interview-challenge"
version = "0.1.0"
description = ""
authors = ["valnog <nogueiravalente@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
django = "^5.1.2"
djangorestframework = "^3.15.2"

bandit = {extras = ["toml"], version = "^1.7.10"}
django-extensions = "^3.2.3"
cpf-cnpj-validate = "^1.4"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"
taskipy = "^1.13.0"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Configuration

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "picpay.settings"
pythonpath = "."
addopts = "-p no:warnings"

[tool.bandit]
confidence = 90
exclude_dirs = ["tests", "migrations"]

[tool.coverage.run]
source = ['interview_challenge']

[tool.ruff]
line-length = 119
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.taskipy.tasks]
run = './manage.py runserver'
lint = 'ruff check ./picpay && ruff check ./picpay --diff'
format = 'ruff check ./picpay --fix && ruff format ./picpay'
bandit = 'bandit -r .'
code_convention = 'task lint && task format'
pre_test = 'task bandit && task code_convention'
test = 'pytest --cov=picpay -vv'
post_test = 'coverage html'